refactor(types): 数学関連の型定義の整理と新しいモジュールの追加
