+++ package/main/src/types/index.ts
@@ -4,4 +4,4 @@ export * from "./dateType";
 export * from "./int";
 export * from "./jokeType";
 export * from "./logicType";
-export * from "./mathType";
+export * from "./math";
