+++ package/main/src/types/math/index.ts
@@ -1,0 +1,6 @@
+export * from "./add";
+export * from "./divide";
+export * from "./modulo";
+export * from "./multiply";
+export * from "./numberToArray";
+export * from "./subtract";
